refactor DittoProtocolAdapter and add ProtocolFactory, ProtocolAdapter and AbstractProtocolAdapter
