+++ Jenkinsfile
@@ -1,7 +1,7 @@
 #!groovy
 node {
   // Need to replace the '%2F' used by Jenkins to deal with / in the path (e.g. story/...)
-  String theBranch = "${env.BRANCH_NAME}".replace('%2F', '-').replace('/','-');
+    String theBranch = "${env.BRANCH_NAME}".replace('%2F', '-').replace('/','-');
   String theMvnRepo = "$WORKSPACE/../feature-repository-${theBranch}";
 
   stage('Checkout') {
@@ -10,7 +10,7 @@ node {
 
   stage('set version to ${theBranch}') {
     withMaven(
-      maven: 'maven-3.5.0',
+      maven: 'maven-3.5.2',
       mavenLocalRepo: theMvnRepo) {
 
       sh "mvn versions:set -DnewVersion=0-${theBranch}-SNAPSHOT"
@@ -19,7 +19,7 @@ node {
 
   stage('Build') {
     withMaven(
-      maven: 'maven-3.5.0',
+      maven: 'maven-3.5.2',
       mavenLocalRepo: theMvnRepo) {
 
       sh "mvn clean deploy javadoc:jar source:jar-no-fork" +
