+++ services/gateway/endpoints/src/main/java/org/eclipse/ditto/services/gateway/endpoints/routes/RootRoute.java
@@ -45,6 +45,7 @@ import org.eclipse.ditto.model.base.headers.DittoHeaders;
 import org.eclipse.ditto.model.base.headers.DittoHeadersBuilder;
 import org.eclipse.ditto.model.base.json.JsonSchemaVersion;
 import org.eclipse.ditto.model.policies.SubjectIssuer;
+import org.eclipse.ditto.protocoladapter.DittoProtocolAdapter;
 import org.eclipse.ditto.services.gateway.endpoints.directives.CorsEnablingDirective;
 import org.eclipse.ditto.services.gateway.endpoints.directives.EncodingEnsuringDirective;
 import org.eclipse.ditto.services.gateway.endpoints.directives.HttpsEnsuringDirective;
@@ -171,7 +172,8 @@ public final class RootRoute {
         thingSearchRoute = new ThingSearchRoute(proxyActor, actorSystem);
         websocketRoute = new WebsocketRoute(streamingActor,
                 config.getInt(ConfigKeys.WEBSOCKET_SUBSCRIBER_BACKPRESSURE),
-                config.getInt(ConfigKeys.WEBSOCKET_PUBLISHER_BACKPRESSURE));
+                config.getInt(ConfigKeys.WEBSOCKET_PUBLISHER_BACKPRESSURE),
+                DittoProtocolAdapter.newInstance());
 
         supportedSchemaVersions = config.getIntList(ConfigKeys.SCHEMA_VERSIONS);
 
@@ -365,8 +367,8 @@ public final class RootRoute {
                                 mapAuthorizationContext(correlationId, wsVersion,
                                         authContextWithPrefixedSubjects,
                                         authContext ->
-                                                websocketRoute.buildWebsocketRoute(correlationId, authContext,
-                                                        wsVersion)
+                                                websocketRoute.buildWebsocketRoute(wsVersion, correlationId,
+                                                        authContext)
                                 )
                         )
                 )
