+++ services/gateway/proxy/src/main/java/org/eclipse/ditto/services/gateway/proxy/actors/AbstractThingProxyActor.java
@@ -106,9 +106,11 @@ public abstract class AbstractThingProxyActor extends AbstractProxyActor {
                 .match(SudoRetrieveThings.class, command -> {
                     getLogger().debug("Got 'SudoRetrieveThings' message, forwarding to the Things Aggregator");
                     if (command.getThingIds().isEmpty()) {
+                        getLogger().debug("Got 'SudoRetrieveThings' message with no ThingIds");
                         notifySender(SudoRetrieveThingsResponse.of(JsonArray.newBuilder().build(),
                                 command.getDittoHeaders()));
                     } else {
+                        getLogger().debug("Got 'SudoRetrieveThings' message, forwarding to the Things Aggregator");
                         thingsAggregator.forward(command, getContext());
                     }
                 })
@@ -141,11 +143,12 @@ public abstract class AbstractThingProxyActor extends AbstractProxyActor {
 
                 /* Thing Commands */
                 .match(RetrieveThings.class, command -> {
-                    getLogger().debug("Got 'RetrieveThings' message, forwarding to the Things Aggregator");
                     if (command.getThingIds().isEmpty()) {
+                        getLogger().debug("Got 'RetrieveThings' message with no ThingIds");
                         notifySender(RetrieveThingsResponse.of(JsonFactory.newArray(),
                                 command.getNamespace().orElse(null), command.getDittoHeaders()));
                     } else {
+                        getLogger().debug("Got 'RetrieveThings' message, forwarding to the Things Aggregator");
                         thingsAggregator.forward(command, getContext());
                     }
                 })
